// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/filtroElectoresOmisosBean.ts
export class FiltroElectoresOmisosBean{
  esquema?: string;
  idProceso: number;
  proceso: string;
  idEleccion: string;
  eleccion: string;
  idAmbito: number;
  idCentroComputo: number;
  centroComputo: string;
  codigoCentroComputo: string;
  tipoReporte?: number;
  tipoReporteActorElectoral?: number;
  acronimo:string;
}